requirements.txt
anthropic
flask
python-docx
python-dotenv
chromadb
pymupdf
sentence-transformers
